Flask-VanillaJs/server/requirements.py3.10._.txt
# 
# python 3.10 compliant
#

# server framework
Flask==3.1.0
# .env parsing
environs==14.1.1
# external api calls
requests
# google client ( for google services )
google-api-python-client==2.154.0
# google auth
google-auth-oauthlib==1.2.1